Add Installed Engines to Additional Load Paths
Although we have a means of adding pre-built JS from engines into our
applications, it doesn't quite behave the same way in some
circumstances. For example, in a situation wherein you want to provide
JavaScript that should be included into the application's webpack and
possibly overridden, the official prescribed way would not work out very
well.

Alternatively, there's a pretty low-effort way of doing this from
engines, by utilizing the `additional_paths` method inside
`Webpacker::Configuration`. To accomplish this, we're iterating over all
`Rails::Engine` subclasses and adding their **app/javascript** directory
to the collection of additional paths.

Here's an example of how this might work in an engine:

```javascript

export default { foo: "bar" }
```

And in a host application:

```javascript
import Engine from "my-cool-engine"

Engine // => { foo: "bar" }
```

While this might not work for everyone, I feel that it removes a lot of
the rigidity and complexity in working with engines using the current
approach.